🐛 isolate test case for global setup
