+++ packages/varmint/__tests__/filebox.test.ts
@@ -1,10 +1,11 @@
+import { spawn } from "node:child_process"
 import * as fs from "node:fs"
 import * as http from "node:http"
 import path from "node:path"
 
 import * as tmp from "tmp"
 
-import { Squirrel, varmintWorkspaceManager } from "../src"
+import { Squirrel } from "../src"
 
 let server: http.Server
 let tempDir: tmp.DirResult
@@ -99,20 +100,17 @@ describe(`Filebox`, () => {
 		}
 	})
 	test(`flushing untouched files`, async () => {
-		function asyncRand() {
-			return new Promise((resolve) =>
-				setTimeout(() => {
-					resolve(Math.floor(Math.random() * 100))
-				}, 1),
-			)
-		}
-		varmintWorkspaceManager.startGlobalTracking()
-		const squirrel = new Squirrel(`read-write`, tempDir.name)
-		const rand = squirrel.add(`rand`, asyncRand)
-		const myRand = rand.for(`my-rand`)
-		await myRand.get()
-
+		const setup = spawn(
+			`node`,
+			[`--experimental-strip-types`, `global-setup.node.ts`, tempDir.name],
+			{
+				stdio: `inherit`,
+				cwd: path.join(import.meta.dirname, `isolation`),
+			},
+		)
+		await new Promise((resolve) => setup.on(`exit`, resolve))
 		fs.mkdirSync(path.join(tempDir.name, `other`))
+		console.log(`tempDir contents:`, fs.readdirSync(tempDir.name))
 		fs.writeFileSync(
 			path.join(tempDir.name, `rand`, `some-random-file.whatever`),
 			`{}`,
@@ -123,7 +121,15 @@ describe(`Filebox`, () => {
 			`my-rand.output.json`,
 			`some-random-file.whatever`,
 		])
-		varmintWorkspaceManager.endGlobalTrackingAndFlushUnusedFiles()
+		const teardown = spawn(
+			`node`,
+			[`--experimental-strip-types`, `global-teardown.node.ts`],
+			{
+				stdio: `inherit`,
+				cwd: path.join(import.meta.dirname, `isolation`),
+			},
+		)
+		await new Promise((resolve) => teardown.on(`exit`, resolve))
 		expect(fs.readdirSync(tempDir.name)).toEqual([`rand`])
 		expect(fs.readdirSync(path.join(tempDir.name, `rand`))).toEqual([
 			`my-rand.input.json`,
