+++ packages/varmint/__tests__/isolation/global-teardown.node.ts
@@ -1,0 +1,3 @@
+import { varmintWorkspaceManager } from "../../src/varmint-workspace-manager.ts"
+
+varmintWorkspaceManager.endGlobalTrackingAndFlushUnusedFiles()
